Feature/java ref impl (#14)
* stuff

* stuff

* reference impl - mildly buggy

* remove unused files

* move stuff around

* why is inserting into DB not working any longer

* write png and ppm

* fix ref impl

* fix ref impl

* avoid writing pngs and svgs to disk

* clean up

* remove logging

* stuff

* updated Graal VM, but not working

* cpp impl (#12)

* cpp impl

* update deps

* models

* stuff

* readme

* add Cargo.lock

* fmt

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/graal vm ref impl (#4)

* graal vm

* graal vm

* graal vm

* gitignore (#5)

* rename stuff (#6)

* parser generator (#7)

* parser generator

* make it compile

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* Feature/images as png and json (#11)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* Feature/graal vm ref impl (#4)

* graal vm

* graal vm

* graal vm

* gitignore (#5)

* rename stuff (#6)

* parser generator (#7)

* parser generator

* make it compile

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* saving stuff

* cpp impl crashes with a segfault - no surprise here

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

---------


* reference impl - mildly buggy

* cpp impl (#12)

* cpp impl

* update deps

* models

* stuff

* readme

* add Cargo.lock

* fmt

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/graal vm ref impl (#4)

* graal vm

* graal vm

* graal vm

* gitignore (#5)

* rename stuff (#6)

* parser generator (#7)

* parser generator

* make it compile

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* Feature/images as png and json (#11)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* Feature/graal vm ref impl (#4)

* graal vm

* graal vm

* graal vm

* gitignore (#5)

* rename stuff (#6)

* parser generator (#7)

* parser generator

* make it compile

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* saving stuff

* cpp impl crashes with a segfault - no surprise here

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

* Feature/images as png and json (#2)

* saving stuff

* add cargo lock
update dependencies

* Feature/resize image (#3)

* resize image thingi

* cargo.lock

* wiki db

* update deps

* saving stuff

* add cargo lock
update dependencies

* Feature/frontend (#9)

* sycamore router

* fmt

* fix html title

* Feature/frontend (#10)

* sycamore router

* fmt

* fix html title

---------


* stuff

* reference impl - mildly buggy

* write png and ppm

---------

+++ rust/common/src/models.rs
@@ -9,7 +9,7 @@ pub const TABLE_WIKIPAGE: &str = "wiki_page";
 
 #[derive(Debug, Clone)]
 pub struct ArticleModel {
-    pub id: i32,
+    pub id: i64,
     pub code: String,
     pub title: String,
     pub description: String,
@@ -33,7 +33,7 @@ pub struct NewImageModel {
 
 #[derive(Debug)]
 pub struct ImageModel {
-    pub id: i32,
+    pub id: i64,
     pub filename: String,
     pub image_as_rgb_png: String,
     pub image_as_json_pixels_array: String,
@@ -43,15 +43,15 @@ pub struct ImageModel {
 
 #[derive(Debug)]
 pub struct NewArt2ImgModel {
-    pub article_id: i32,
-    pub image_id: i32,
+    pub article_id: i64,
+    pub image_id: i64,
 }
 
 #[derive(Debug)]
 pub struct Art2ImgModel {
-    pub id: i32,
-    pub article_id: i32,
-    pub image_id: i32,
+    pub id: i64,
+    pub article_id: i64,
+    pub image_id: i64,
 }
 
 #[derive(Debug)]
@@ -61,7 +61,7 @@ pub struct NewResolutionModel {
 
 #[derive(Debug)]
 pub struct ResolutionModel {
-    pub id: i32,
+    pub id: i64,
     pub resolution: String,
 }
 
@@ -112,7 +112,7 @@ impl From<&Row> for ResolutionModel {
 
 #[derive(Debug)]
 pub struct NewWikipageModel {
-    pub id: i32,
+    pub id: i64,
     pub title: Option<String>,
     pub ns: Option<String>,
     pub redirect: Option<String>,
@@ -127,8 +127,8 @@ pub struct NewWikipageModel {
 
 #[derive(Debug)]
 pub struct WikipageModel {
-    pub dbid: i32,
-    pub id: i32,
+    pub dbid: i64,
+    pub id: i64,
     pub title: Option<String>,
     pub ns: Option<String>,
     pub redirect: Option<String>,
